完成所有课程中的内容

+++ 保卫萝卜demo/Assets/Game/Scripts/Application/StaticData/BulletInfo.cs
@@ -7,5 +7,5 @@ public class BulletInfo
 	public int BulletID;
 	public string PrefabName;
 	public float BaseSpeed;
-	public float BaseAttack;
+	public int BaseAttack;
 }
